feat: Update code and tests to insert content after comment placeholders
- Modify `insertContentAfterComments` function to insert content after comment placeholders in the template.
- Update test cases to reflect the changes in functionality.
- Adjust expected output in tests to include content after comments.
