+++ src/lib/file.ts
@@ -53,20 +53,22 @@ export const readMdFiles = async (dirPath: string): Promise<MdFiles> => {
 };
 
 /**
- * Replaces comment placeholders in a template with corresponding content from insertDatas.
+ * Inserts content from insertDatas after comment placeholders in a template.
  *
  * @param {string} template - The template string containing comment placeholders.
  * @param {MdFiles} insertDatas - An array of objects containing markdown content keyed by filenames.
- * @returns {string} The template with comment placeholders replaced by actual content.
+ * @returns {string} The template with content inserted after comment placeholders.
  */
-export const replaceCommentToMd = (
+export const insertContentAfterComments = (
   template: string,
   insertDatas: MdFiles
 ): string => {
   return template.replace(
     /<!--\s*([a-zA-Z0-9_.-]+)\s*-->/g,
-    (match, fileName) =>
-      insertDatas.find((md) => md[fileName])?.[fileName] || match
+    (match, fileName) => {
+      const content = insertDatas.find((md) => md[fileName])?.[fileName];
+      return content ? `${match}\n${content}` : match;
+    }
   );
 };
